pyproject.toml
[tool.poetry]
name = "hw27"
version = "0.1.0"
description = ""
authors = ["Ivan <ifucin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
python-dateutil = "^2.8.2"
pytest-factoryboy = "^2.5.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
